XF3-207: Add in MIPs source to ccsp-p-and-m
Reason for change: The Makefiles wer changed in the XB3 build that
require the build to use either PC ro ARM build options. Since
the Pace XF3 is neither of those, we need to add in MIPs source
options

Test Procedure: Verified that the build completes.

Risks: Medium. The changes are not likley to run properly, but the
build should complete without an issue.

Signed-off-by: scott_angel@cable.comcast.com
Change-Id: I5c655e1c35505464065f4f47f1eba78e1d8ee2a8