RDKB-9577: User able to browse after activation for 2 minutes
Reason for change: After activation, user is able to browse for a short period of time
In such scenario as device is whitelisted by DNS, it is possible that a client can browse as internal LAN services are still in the process of restarting.
Change made to firewall restart earlier so that any query on port 53 will be redirected to 10.0.0.1

Test Procedure:
Verified fix by doing factory reset
Simulated a case where unit will be in preactivated state. In that configuration iptable rule was added properly after getting 204 response.

Change-Id: I90b0740572006d4a03082170a1ea974150b5957b
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>