RDKB-11354 : Unknown Reboot Reason for GUI Reboot
Reason for change:
Setting of reboot reason was commented as part of restore configuration fix,
as committing the syscfg db caused overwriting of replaced db.
So updated confPhp script to set a parameter restore_reboot and
added check in the code to not to commit syscfg if it is a restore reboot.

Test Procedure:
Test 1 : Reboot device from GUI. Verify the last reboot reason once the device is up
Test 2 : Save and Restore of user Configuration from GUI.
Risks: None
Signed-off-by: Dhanya_Ravindran2@comcast.com

Change-Id: I4f04df68dfa20214bd87d6fa246e3444965c8668