RDKB-14923 : [Master] Fail to associate LAN Ethernet Port to XHN.
Reason for change:
As part of RDKB-10632 this ticket, we have made error handling for ioctl calls for invalid interface.
Problamatic case we are trying to query invalid sw_4 interface name. so that it is failed to proceed further. so needs to avoid sw_ prefix port names from ioctl calls.

Test Procedure:
1. Able to do below dmcli command,
   dmcli eRT setv Device.Bridging.Bridge.2.Port.2.Enable bool true
2. Able to associate LAN Ethernet Port to XHN via WEBUI page.
3. Ensure PAM bridging functionality should not break.

Change-Id: Id1f8b123341a3020b5ccb28d79fe2e8472e5fc62
Risks: None
Signed-off-by: Lakshminarayanan Shenbagaraj Lakshminarayanan_Shenbagaraj@comcast.com