RDKB-15514: Restructure UI to have TFTP dropdown
Reason for change: As part of RDKB-15051, the security team required that we
do not allow the server string to be visible in the UI. Now the UI has a
selectable drop down with TFTP server indexes.

Test Procedure: Verify that the TFTP server is set and saved propery

Risks: Med.

Change-Id: I199c3aa27f1c7b1742e59ae1266cd240a28597f9
Signed-off-by: scott_angel@cable.comcast.com