TCCBR-1960: Fix PAM crash while setting True Static IP
Reason for change: The original code for setting the static IP
used AnscSpawnTask which defaults to a 32k stack size. In this
case, the stack size was too small due to the number of function
calls required in the spawned thread.

Test Procedure: Verify that PAM does not create when the cliconfig.txt
file is set for static IP

Risks: Low

Change-Id: I0f7bd762823202c01a73cc78065ecfb3d87e8c60
Signed-off-by: scott_angel@cable.comcast.com