XF3-2959: [CFG3] PD configs missing in CFG3 dibbler server Configs
Reason for change: The pd-class block in the dibbler server.conf file is
missing.  Implemented the required logics to genrate the pd-classs block
in the dibbler server.conf file during bootup.  The dhcpv6Prefix.patch
file was also folded to this change based on review commnets.

Test Procedure: Make sure the pd-classs block is in the dibbler server.conf
file.

Risks: Low

Change-Id: I851a5a228f0a7805c661cf81a4db57ff4ccaaac0
Signed-off-by: alui <Alvin_Lui2@cable.comcast.com>