RDKCOM-2327: RDKBDEV-564: fix misuse of "\0"
Reason for change:
"\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).
Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.
Test Procedure: Sanity.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit a1a5a61c7650f0f4dfb1ccb270a1a93bfd205a91)
Change-Id: I893f93c78a8ff1bda9313ca8b9a0ff53018eeb1e