RDKBDEV-563: ensure busybox ps command is used
Reason for change:
ensure busybox ps command is used (and not the procps version)

For calls to ps using the busybox compatible syntax, ensure that the
busybox version of ps is actually used (ie run "busybox ps" instead
of "ps").

Note that the busybox version of ps is assumed for embedded targets
(source-arm, source-mips, etc) and the procps version of ps is
assumed for source-pc.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Ib39ea89311fa30a861bb6db7dc80ef3237acc2d2