RDKBDEV-562: use busybox compatible syntax consistently when calling "ps"
Reason for change:
The minimal busybox and full-featured procps versions "ps" do not
support the same command line options. The code currently seems to
contain a mixture of busybox and procps compatible syntax:

  procps  : ps -A | grep XXX
  busybox : ps | grep XXX | grep -v grep

In sources intended to run on embedded targets (ie source-arm,
source-mips, etc but not source-pc) replace the procps compatible
commands with the busybox equivalents.

Note that procps compatible syntax does not work on targets using
busybox ps, so it seems that these functions are not used / tested.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I799502069153bfe069e668b97cd7b35481bde4b9