RDKCOM-2101: fix incorrect definition of MACADDR_SZ in cosa_ethernet_apis.h
Reason for change: The size of a buffer needed to contain a MAC address in
string format is 18, not 17. Fix the definition and therefore the size of the
MacAddress buffer in the COSA_DML_ASSOCDEV_INFO structure.
Drop workaround for the incorrect definition from CosaEthPortGetAssocDevices().
Test Procedure: Build and Sanity test
Risks: None
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Iccfd9e2b91dba6c01a32efb74e4bb860974965bb