RDKBDEV-376: fix incorrect definition of MACADDR_SZ in cosa_ethernet_apis.h
Reason for change: The size of a buffer needed to contain a MAC address in
string format is 18, not 17. Fix the definition and therefore the size of the
MacAddress buffer in the COSA_DML_ASSOCDEV_INFO structure.
Drop workaround for the incorrect definition from CosaEthPortGetAssocDevices().
Test Procedure: Build and Sanity test
Risks: None

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Iedf70e509ace93644cc92836c23cfa9a70e0f64a