RDKBDEV-927: fix invalid string passed to Utopia_RawSet()
Reason for change:
fix invalid string passed to Utopia_RawSet() in CosaDmlRaIfSetCfg()
Ensure that the strings passed to Utopia_RawSet() in
CosaDmlRaIfSetCfg() are correctly nul terminated (it's not safe to
construct a string by writing a single char to the first location in
a buffer and hoping that the second location is a nul...).

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I35dcee4e9d424f0d25e8b326a0796eb0199991a6