RDKCOM-2868: RDKBDEV-927: fix invalid string passed to Utopia_RawSet()
Reason for change:
fix invalid string passed to Utopia_RawSet() in CosaDmlRaIfSetCfg()
Ensure that the strings passed to Utopia_RawSet() in
CosaDmlRaIfSetCfg() are correctly nul terminated (it's not safe to
construct a string by writing a single char to the first location in
a buffer and hoping that the second location is a nul...).
Test Procedure: Sanity.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit 4e3309d1b248bcf360337b7fead420ffbc0da10e)
Change-Id: I406089c187ad055431dbe38b21bf7e5228cb851f