RDKBDEV-903: drop support for _COSA_DRG_CNS_ and _COSA_DRG_TPG_
Reason for change:
Drop support for obsolete _COSA_DRG_CNS_ and _COSA_DRG_TPG_ platforms
These are legacy Cisco platforms which are not supported or buildable
in RDKB. They predate the introduction of the RDKB HAL APIs so the
code isn't generally useful to keep around, even as a reference.

  for file in $(find . -type f -name '*.[ch]') ; do unifdef -U_COSA_DRG_TPG_ -U_COSA_DRG_CNS_ -o $file $file ; done

drop further references to _COSA_DRG_CNS_ and _COSA_DRG_TPG_ platforms
As a follow up to the previous commit to drop support for legacy
platforms, remove references to _COSA_DRG_TPG_ etc which are not
automatically removed by unifdef.

Test Procedure: Build and Sanity.
Risks None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I06e0cead400fe0ad63381fe1a523e31bb779afe5