RDKBDEV-710 DDNS: timeout error code is seen while setting DDNS data models
Reason for change: On analysis it was found that when the data model parameters
are changed then service_dynamic_dns.sh is getting invoked in the foreground and
the vsystem() command is waiting for the script to finish running. This can cause
timeout errors. Running the script in the background makes vsystem() or v_secure_system() 
to not wait for the script to finish.

Test Procedure: Updated in JIRA

Risks: None

Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>
Change-Id: I263c4b8960020059aa17d14e38faa61f8f4dbe92