RDKCOM-2566: RDKBDEV-710 DDNS: timeout error code is seen while setting DDNS...
Reason for change: On analysis it was found that when the data model parameters
are changed thenservice_dynamic_dns.shis getting invoked in the foreground and
thevsystem()command is waiting for the script to finish running. This can cause
timeout errors. Running the script in the background makesvsystem()orv_secure_system()
to not wait for the script to finish.
Test Procedure: Updated in JIRA
Risks: None
Signed-off-by: Mohamed Shaikh <mohamed.shaikh@t-systems.com>

(cherry picked from commit 4c0b98af094cd53d878872e521b979bb8616844a)
Change-Id: If914bbbdb12d4a765369544d37b093656ce3bf0d