RDKBDEV-1205: fix output buffer size checks for X_CISCO_COM_Password
Reason for change:
fix output buffer size checks for X_CISCO_COM_Password in User_GetParamStringValue()
The string which will be returned may be the password or the hashed
password. Check that the output buffer is big enough for the worst
case (ie maximum size of the hashed password).

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I0212069b7a9e5bec8e0a050f9a65ca21f6e204cc